Return 404 Not Found when a non-existing entity is requested

+++ src/ProtoR.Application/Category/DeleteCategoryCommandHandler.cs
@@ -4,6 +4,7 @@ namespace ProtoR.Application.Category
     using System.Threading.Tasks;
     using MediatR;
     using ProtoR.Domain.CategoryAggregate;
+    using ProtoR.Domain.Exceptions;
     using ProtoR.Domain.SeedWork;
 
     public class DeleteCategoryCommandHandler : AsyncRequestHandler<DeleteCategoryCommand>
@@ -21,6 +22,13 @@ namespace ProtoR.Application.Category
 
         protected override async Task Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
+            var category = await this.categoryRepository.GetById(request.CategoryId);
+
+            if (category == null)
+            {
+                throw new EntityNotFoundException<Category>(request.CategoryId);
+            }
+
             // TODO groups assigned to deleted category
             await this.categoryRepository.Delete(request.CategoryId);
             await this.unitOfWork.SaveChanges();
