+++ src/ProtoR.Application/Client/DeleteClientCommandHandler.cs
@@ -4,6 +4,7 @@ namespace ProtoR.Application.Client
     using System.Threading.Tasks;
     using MediatR;
     using ProtoR.Domain.ClientAggregate;
+    using ProtoR.Domain.Exceptions;
     using ProtoR.Domain.SeedWork;
 
     public class DeleteClientCommandHandler : AsyncRequestHandler<DeleteClientCommand>
@@ -21,6 +22,13 @@ namespace ProtoR.Application.Client
 
         protected override async Task Handle(DeleteClientCommand request, CancellationToken cancellationToken)
         {
+            var client = await this.clientRepository.GetById(request.ClientId);
+
+            if (client == null)
+            {
+                throw new EntityNotFoundException<Client>(request.ClientId);
+            }
+
             await this.clientRepository.Delete(request.ClientId);
             await this.unitOfWork.SaveChanges();
         }
